refactor storage state
