+++ frontend/testing/playwright/helpers/DesignerApi.ts
@@ -1,12 +1,17 @@
 import { APIRequestContext, APIResponse } from '@playwright/test';
 import { Environment, StudioEnvironment } from './StudioEnvironment';
+import type { Cookie, StorageState } from '../types/StorageState';
 
 export class DesignerApi extends StudioEnvironment {
   constructor(environment?: Environment) {
     super(environment);
   }
 
-  public async createApp(request: APIRequestContext, xsrfToken: string): Promise<APIResponse> {
+  public async createApp(
+    request: APIRequestContext,
+    storageState: StorageState,
+  ): Promise<APIResponse> {
+    const xsrfToken: string = this.getXsrfTokenFromStorageState(storageState);
     const response = await request.post(
       `/designer/api/repos/create-app?org=${this.org}&repository=${this.app}&datamodellingPreference=1`,
       {
@@ -17,5 +22,13 @@ export class DesignerApi extends StudioEnvironment {
       },
     );
     return response;
+  }
+
+  private getXsrfTokenFromStorageState(storageState: StorageState): string {
+    const fs = require('fs');
+    const jsonData = fs.readFileSync(storageState, 'utf-8');
+    const formattedStorageState: StorageState = JSON.parse(jsonData);
+    return formattedStorageState.cookies.find((cookie: Cookie) => cookie.name === 'XSRF-TOKEN')
+      .value;
   }
 }
