--- frontend/testing/playwright/types/Cookies.ts
@@ -1,15 +1,0 @@
-export interface Cookie {
-  name: string;
-  value: string;
-  domain: string;
-  path: string;
-  expires: number;
-  httpOnly: boolean;
-  secure: boolean;
-  sameSite: string;
-}
-
-export interface CookiesObject {
-  cookies: Cookie[];
-  origins: string[];
-}
