+++ frontend/testing/playwright/types/StorageState.ts
@@ -1,0 +1,8 @@
+export interface Cookie {
+  name: string;
+  value: string;
+}
+
+export interface StorageState {
+  cookies: Cookie[];
+}
